requirements.txt
pydantic
sqlmodel
python-docx
openpyxl
PyPDF2
typer
pytest
